Checkpoint towards using package-bundler
This would give us two things: publishing as ESM in addition to CJS, and
the ability to have some symbols come from "deep imports" that aren't
loaded automatically. It replaces some hacks like runtime requires in
src/plugin/index.ts, and adds things like "you don't need to load
express unless you actually import from the standalone directory".

A directory with an index.ts in it is an entry point. The main rule is:
you can import arbitrarily within your own entry point, but you can't do
a "deep" import of a specific file or directory inside another entry
point. Note that this means that any code that needs to be accessible
from multiple entry points but not externally importable should be in an
entry point whose name contains "internal" so that folks understand they
shouldn't rely on its existence.

This is coming close to working. Some issues I'm running into include:

- TS directive emission isn't working because of the change to make
  packages/server/tsconfig.json only cover package.json. But maybe that
  can be reverted somehow.
- How do we actually enforce that bad deep imports don't happen?
- Is there still any point to ever running tsc emitting output? (Like
  for Jest maybe?) Should we ditch dist entirely? Not sure how though.
- We're using the "copy and rewrite package.json into output dir"
  approach because that way the deep imports are
  `@apollo/server/standalone` not `@apollo/server/bundled/standalone`. I
  already found out how to get codesandbox CI to publish from the right
  directory but we have to do the same for changesets probably.
- I did some refactors to make ApolloServer.ts use runtime import() to
  pull in the default-installed plugins but in CJS mode that still seems
  to include the plugin's code in the CJS file --- it won't actually be
  run (including its import of protobuf stuff) until needed at least.

Let's figure out more next week.
