+++ .codesandbox/ci.json
@@ -1,10 +1,14 @@
 {
-  "buildCommand": "compile",
+  "buildCommand": "bundle",
   "installCommand": "install-with-npm-8.5",
   "packages": [
     "packages/usage-reporting-protobuf",
     "packages/server"
   ],
+  "publishDirectory": {
+    "@apollo/server": "packages/server/bundled",
+    "@apollo/usage-reporting-protobuf": "packages/usage-reporting-protobuf"
+  },
   "sandboxes": ["apollo-server-typescript-3opde","apollo-server"],
   "node": "16"
 }
