+++ package.json
@@ -4,7 +4,8 @@
   "license": "MIT",
   "repository": "github:apollographql/apollo-server",
   "scripts": {
-    "clean": "git clean -dfqX -- ./node_modules **/{dist,node_modules}/ **/tsconfig*tsbuildinfo",
+    "bundle": "cd packages/server && package-bundler --platform node --outDir ./bundled --copyPackageJson --rewritePackageJson",
+    "clean": "git clean -dfqX -- ./node_modules **/{bundled,dist,node_modules}/ **/tsconfig*tsbuildinfo",
     "compile": "tsc --build tsconfig.build.json",
     "compile:clean": "tsc --build tsconfig.build.json --clean",
     "watch": "tsc --build tsconfig.build.json --watch",
@@ -33,6 +34,7 @@
   "devDependencies": {
     "@apollo/client": "3.6.7",
     "@apollo/utils.createhash": "1.1.0",
+    "@better-builds/package-bundler": "^1.2.8",
     "@graphql-codegen/cli": "2.6.2",
     "@graphql-codegen/typescript-operations": "2.4.2",
     "@graphql-tools/mock": "8.6.12",
@@ -77,6 +79,7 @@
     "jest-junit": "13.2.0",
     "jest-mock-random": "1.1.1",
     "lodash.sumby": "4.6.0",
+    "lru-cache": "7.10.1",
     "nock": "13.2.6",
     "node-fetch": "2.6.7",
     "prettier": "2.6.2",
