+++ packages/server/.npmignore
@@ -1,7 +1,7 @@
 *
 !src/**/*
 src/**/__tests__/**
-!dist/**/*
-dist/**/__tests__/**
+!bundled/**/*
+bundled/**/__tests__/**
 !package.json
 !README.md
