+++ packages/server/src/ApolloServer.ts
@@ -1,5 +1,11 @@
 import { isNodeLike } from '@apollo/utils.isnodelike';
+import {
+  InMemoryLRUCache,
+  PrefixingKeyValueCache,
+  type KeyValueCache,
+} from '@apollo/utils.keyvaluecache';
 import type { Logger } from '@apollo/utils.logger';
+import type { WithRequired } from '@apollo/utils.withrequired';
 import { addMocksToSchema } from '@graphql-tools/mock';
 import { makeExecutableSchema } from '@graphql-tools/schema';
 import resolvable, { Resolvable } from '@josephg/resolvable';
@@ -15,46 +21,38 @@ import {
   print,
   ValidationContext,
 } from 'graphql';
-import {
-  type KeyValueCache,
-  InMemoryLRUCache,
-  PrefixingKeyValueCache,
-} from '@apollo/utils.keyvaluecache';
 import loglevel from 'loglevel';
 import Negotiator from 'negotiator';
 import * as uuid from 'uuid';
-import { newCachePolicy } from './cachePolicy';
 import { determineApolloConfig } from './determineApolloConfig';
 import { BadRequestError, ensureError, formatApolloErrors } from './errors';
 import type {
+  ApolloConfig,
+  ApolloServerOptions,
   ApolloServerPlugin,
   BaseContext,
+  ContextThunk,
+  DocumentStore,
   GraphQLExecutor,
   GraphQLRequest,
   GraphQLResponse,
-  GraphQLServerListener,
   GraphQLServerContext,
+  GraphQLServerListener,
+  HTTPGraphQLHead,
   HTTPGraphQLRequest,
   HTTPGraphQLResponse,
   LandingPage,
-  ApolloConfig,
-  ApolloServerOptions,
-  DocumentStore,
   PersistedQueryOptions,
-  HTTPGraphQLHead,
-  ContextThunk,
 } from './externalTypes';
+import type { ApolloServerOptionsWithStaticSchema } from './externalTypes/constructor';
 import { runPotentiallyBatchedHttpQuery } from './httpBatching';
-import { InternalPluginId, pluginIsInternal } from './internalPlugin';
 import {
-  ApolloServerPluginCacheControl,
-  ApolloServerPluginInlineTrace,
-  ApolloServerPluginLandingPageLocalDefault,
-  ApolloServerPluginLandingPageProductionDefault,
-  ApolloServerPluginSchemaReporting,
-  ApolloServerPluginSchemaReportingOptions,
-  ApolloServerPluginUsageReporting,
-} from './plugin';
+  HeaderMap,
+  InternalPluginId,
+  isImplicitlyInstallablePlugin,
+  newCachePolicy,
+  pluginIsInternal,
+} from './internal';
 import {
   preventCsrf,
   recommendedCsrfPreventionRequestHeaders,
@@ -63,14 +61,11 @@ import { APQ_CACHE_PREFIX, processGraphQLRequest } from './requestPipeline';
 import {
   badMethodErrorMessage,
   cloneObject,
-  HeaderMap,
   newHTTPGraphQLHead,
   prettyJSONStringify,
 } from './runHttpQuery';
-import { SchemaManager } from './utils/schemaManager';
 import { isDefined } from './utils/isDefined';
-import type { WithRequired } from '@apollo/utils.withrequired';
-import type { ApolloServerOptionsWithStaticSchema } from './externalTypes/constructor';
+import { SchemaManager } from './utils/schemaManager';
 
 const NoIntrospection = (context: ValidationContext) => ({
   Field(node: FieldDefinitionNode) {
@@ -395,7 +390,7 @@ export class ApolloServer<TContext extends BaseContext = BaseContext> {
     try {
       // Now that you can't call addPlugin any more, add default plugins like
       // usage reporting if they're not already added.
-      this.addDefaultPlugins();
+      await this.addDefaultPlugins();
 
       const toDispose: (() => Promise<void>)[] = [];
       const executor = await schemaManager.start();
@@ -853,7 +848,7 @@ export class ApolloServer<TContext extends BaseContext = BaseContext> {
     this.internals.state = { phase: 'stopped', stopError: null };
   }
 
-  private addDefaultPlugins() {
+  private async addDefaultPlugins() {
     const { plugins, apolloConfig, logger, nodeEnv } = this.internals;
     const isDev = nodeEnv !== 'production';
 
@@ -865,6 +860,9 @@ export class ApolloServer<TContext extends BaseContext = BaseContext> {
     // Special case: cache control is on unless you explicitly disable it.
     {
       if (!alreadyHavePluginWithInternalId('CacheControl')) {
+        const { ApolloServerPluginCacheControl } = await import(
+          './plugin/cacheControl'
+        );
         plugins.push(ApolloServerPluginCacheControl());
       }
     }
@@ -879,6 +877,9 @@ export class ApolloServer<TContext extends BaseContext = BaseContext> {
           // Keep this plugin first so it wraps everything. (Unfortunately despite
           // the fact that the person who wrote this line also was the original
           // author of the comment above in #1105, they don't quite understand why this was important.)
+          const { ApolloServerPluginUsageReporting } = await import(
+            './plugin/usageReporting'
+          );
           plugins.unshift(ApolloServerPluginUsageReporting());
         } else {
           logger.warn(
@@ -898,8 +899,10 @@ export class ApolloServer<TContext extends BaseContext = BaseContext> {
       const enabledViaEnvVar = process.env.APOLLO_SCHEMA_REPORTING === 'true';
       if (!alreadyHavePlugin && enabledViaEnvVar) {
         if (apolloConfig.key) {
-          const options: ApolloServerPluginSchemaReportingOptions = {};
-          plugins.push(ApolloServerPluginSchemaReporting(options));
+          const { ApolloServerPluginSchemaReporting } = await import(
+            './plugin/schemaReporting'
+          );
+          plugins.push(ApolloServerPluginSchemaReporting());
         } else {
           throw new Error(
             "You've enabled schema reporting by setting the APOLLO_SCHEMA_REPORTING " +
@@ -921,6 +924,9 @@ export class ApolloServer<TContext extends BaseContext = BaseContext> {
         // federated" mode.  (This is slightly different than the
         // pre-ApolloServerPluginInlineTrace where we would also avoid doing
         // this if an API key was configured and log a warning.)
+        const { ApolloServerPluginInlineTrace } = await import(
+          './plugin/inlineTrace'
+        );
         plugins.push(
           ApolloServerPluginInlineTrace({ __onlyIfSchemaIsFederated: true }),
         );
@@ -944,6 +950,10 @@ export class ApolloServer<TContext extends BaseContext = BaseContext> {
       'LandingPageDisabled',
     );
     if (!alreadyHavePlugin) {
+      const {
+        ApolloServerPluginLandingPageLocalDefault,
+        ApolloServerPluginLandingPageProductionDefault,
+      } = await import('./plugin/landingPage/default');
       const plugin: ApolloServerPlugin<TContext> = isDev
         ? ApolloServerPluginLandingPageLocalDefault()
         : ApolloServerPluginLandingPageProductionDefault();
@@ -1225,15 +1235,4 @@ export async function internalExecuteOperation<TContext extends BaseContext>({
     throw new Error('Internal server error');
   }
   return requestContext.response;
-}
-
-export type ImplicitlyInstallablePlugin<TContext extends BaseContext> =
-  ApolloServerPlugin<TContext> & {
-    __internal_installed_implicitly__: boolean;
-  };
-
-export function isImplicitlyInstallablePlugin<TContext extends BaseContext>(
-  p: ApolloServerPlugin<TContext>,
-): p is ImplicitlyInstallablePlugin<TContext> {
-  return '__internal_installed_implicitly__' in p;
 }
