+++ packages/server/src/__tests__/integration/httpServerTests.ts
@@ -27,16 +27,16 @@ import type {
   CreateServerForIntegrationTests,
   CreateServerForIntegrationTestsOptions,
 } from '.';
-import {
+import type {
   ApolloServer,
   ApolloServerOptions,
-  ApolloServerPluginCacheControl,
-  ApolloServerPluginCacheControlDisabled,
   BaseContext,
   GraphQLRequestContext,
   GraphQLRequestListener,
   PersistedQueryOptions,
 } from '../..';
+import { ApolloServerPluginCacheControlDisabled } from '../..';
+import { ApolloServerPluginCacheControl } from '../../plugin/cacheControl';
 import { PersistedQueryNotFoundError } from '../../errors';
 
 const QueryRootType = new GraphQLObjectType({
