+++ packages/server/src/index.ts
@@ -1,3 +1,4 @@
+// TODO(AS4): Consider moving to a deep export
 export {
   SyntaxError,
   ValidationError,
@@ -7,12 +8,18 @@ export {
 } from './errors';
 
 export { ApolloServer } from './ApolloServer';
-export { expressMiddleware } from './express';
-export { startStandaloneServer } from './standalone';
 
 export * from './externalTypes';
 
-export * from './plugin';
+export {
+  ApolloServerPluginCacheControlDisabled,
+  ApolloServerPluginInlineTraceDisabled,
+  ApolloServerPluginLandingPageDisabled,
+  ApolloServerPluginUsageReportingDisabled,
+} from './disabledPlugin';
+
+// TODO(AS4): Consider moving to deep exports
+export { expressMiddleware } from './express';
 
 // TODO(AS4): We do need to export usage reporting protobuf for the purpose of
 // gateway, but we really don't want that to be in top level. Look into this as
