+++ packages/server/src/plugin/internal/computeCoreSchemaHash.ts
@@ -1,0 +1,5 @@
+import { createHash } from '@apollo/utils.createhash';
+
+export function computeCoreSchemaHash(schema: string): string {
+  return createHash('sha256').update(schema).digest('hex');
+}
