+++ packages/server/src/plugin/landingPage/graphqlPlayground/index.ts
@@ -9,7 +9,7 @@ import type {
   ApolloServerPlugin,
   BaseContext,
   GraphQLServerListener,
-} from '../../../externalTypes';
+} from '../../..';
 
 // This specifies the React version of our fork of GraphQL Playground,
 // `@apollographql/graphql-playground-react`.  It is related to, but not to
