+++ packages/server/src/plugin/usageReporting/options.ts
@@ -4,7 +4,7 @@ import type {
   GraphQLRequestContext,
   GraphQLRequestContextWillSendResponse,
   BaseContext,
-} from '../../externalTypes';
+} from '../..';
 import type { Logger } from '@apollo/utils.logger';
 import type { Trace } from '@apollo/usage-reporting-protobuf';
 import type { Fetcher } from '@apollo/utils.fetcher';
