+++ packages/server/src/plugin/usageReporting/plugin.ts
@@ -1,15 +1,17 @@
 import { Report, ReportHeader, Trace } from '@apollo/usage-reporting-protobuf';
 import type { Fetcher, FetcherResponse } from '@apollo/utils.fetcher';
 import {
-  usageReportingSignature,
   calculateReferencedFieldsByType,
   ReferencedFieldsByType,
+  usageReportingSignature,
 } from '@apollo/utils.usagereporting';
 import retry from 'async-retry';
+import { readFileSync } from 'fs';
 import { GraphQLSchema, printSchema } from 'graphql';
 import type LRUCache from 'lru-cache';
 import fetch from 'node-fetch';
 import os from 'os';
+import { join } from 'path';
 import { gzip } from 'zlib';
 import type {
   BaseContext,
@@ -17,13 +19,15 @@ import type {
   GraphQLRequestContextDidResolveOperation,
   GraphQLRequestContextWillSendResponse,
   GraphQLRequestListener,
-  GraphQLServerListener,
   GraphQLServerContext,
-} from '../../externalTypes';
-import type { InternalApolloServerPlugin } from '../../internalPlugin';
-import type { HeaderMap } from '../../runHttpQuery';
-import { computeCoreSchemaHash } from '../schemaReporting';
-import { dateToProtoTimestamp, TraceTreeBuilder } from '../traceTreeBuilder';
+  GraphQLServerListener,
+} from '../..';
+import type { HeaderMap, InternalApolloServerPlugin } from '../../internal';
+import {
+  computeCoreSchemaHash,
+  dateToProtoTimestamp,
+  TraceTreeBuilder,
+} from '../internal';
 import { defaultSendOperationsAsTrace } from './defaultSendOperationsAsTrace';
 import {
   createOperationDerivedDataCache,
@@ -36,8 +40,6 @@ import type {
 } from './options';
 import { OurReport } from './stats';
 import { makeTraceDetails } from './traceDetails';
-import { readFileSync } from 'fs';
-import { join } from 'path';
 
 const reportHeaderDefaults = {
   hostname: os.hostname(),
